Split more poll tests
